claimer don't need an account, split light did creation
