+++ docs/sdk/1-workshop/04-claimer/01-did.md
@@ -6,7 +6,7 @@ title: DID
 import CodeBlock from '@theme/CodeBlock';
 import SnippetBlock from '../../../../src/components/SnippetBlock';
 import GenerateKeypairs from '!!raw-loader!../../../../code-examples/workshop/claimer/generateKeypairs.js';
-import GetLightDid from '!!raw-loader!../../../../code-examples/workshop/claimer/getLightDid.js';
+import GenerateLightDid from '!!raw-loader!../../../../code-examples/workshop/claimer/generateLightDid.js';
 import Index from '!!raw-loader!../../../../code-examples/workshop/claimer/index.js';
 
 Time to make a light DID using the previously created account for the <span class="label-role claimer">Claimer</span>.
@@ -18,50 +18,40 @@ Remember light DIDs can:
 
 Take a look at our [DID documentation](/docs/sdk/core-feature/did) if you want to learn more about DIDs and the difference between their light and full versions.
 
-## generateKeypairs
+## Generate Keys
 
-As before we'll first create our signing and encryption `keypairs`. We'll use an inbound keystore to generate them.
+Similar to the attester, the claimer too has to setup the DID keys.
 Create a file `claimer/generateKeypairs.js` and copy the code below.
 
 <CodeBlock className="language-js" title="claimer/generateKeypairs.js">
   {GenerateKeypairs}
 </CodeBlock>
 
-## getFullDid
+The claimer only needs an authentication key and an encryption key.
+Here the keys are both derived from the same seed, but they could also have two different seeds.
 
-Once our `keypairs` are generated we can create our light DID. Because it's off-chain
-we can just create the DID object everytime, no need to resolve. But we'll still accept
-`didUri` and prompt to save it in `.env` for our reference.
+## Generate Light DID
 
-<CodeBlock className="language-js" title="claimer/getLightDid.js">
-  {GetLightDid}
+Once our `keypairs` are generated we can create our light DID.
+Because it's off-chain we can just create the DID object every time, we don't need to resolve them before using it.
+But we'll still accept `didUri` and prompt to save it in `.env` for our reference.
+
+<CodeBlock className="language-js" title="claimer/generateLightDid.js">
+  {GenerateLightDid}
 </CodeBlock>
 
-## Claimer Index
-
-Ok let's update our `claimer/index.js`. We bring in the `CLAIMER_DID_URI` form `.env` and our `getLightDid` function.
-Finally we'll return the `keystore`, `account` and `lightDid` for use elsewhere later.
-
-<SnippetBlock
-  title="claimer/index.js"
-  className="language-js"
-  snippets='[
-    [0, 17],
-    [20,31],
-    "initialize();"
-  ]'
->
-  {Index}
-</SnippetBlock>
+After everything is initialized, we create a mnemonic that will be used to create the light DID.
+As you may have noticed the claimer doesn't have a `account`.
+The claimer doesn't need to hold funds and also doesn't need a blockchain account.
 
 ## Run
 
 ```bash
-node ./claimer/index.js
+node ./claimer/generateKeypairs.js
 ```
 
-Your output will provide you with `CLAIMER_DID_URI`. Be sure to save it in your `.env`
-file, it should now look similar to this.
+Your output will provide you with `CLAIMER_MNEMONIC` and `CLAIMER_DID_URI`.
+Be sure to save it in your `.env` file, it should now look similar to this.
 
 ```env title=".env"
 WSS_ADDRESS=wss://peregrine.kilt.io/parachain-public-ws
@@ -71,7 +61,6 @@ ATTESTER_ADDRESS=5CUoo2vAegeaZHPNdxZyuMe...
 ATTESTER_DID_URI=did:kilt:4pjUYTbttjJHqT...
 
 CLAIMER_MNEMONIC="gold upset segment cak...
-CLAIMER_ADDRESS=5CUoo2vAegeaZHPNdxZyuMes...
 CLAIMER_DID_URI=did:kilt:light:004tTDugL...
 ```
